+++ src/gretapy/ds/__init__.py
@@ -1,3 +1,3 @@
-from ._db import read_db, read_metrics
+from ._db import read_db, read_imaginary_metrics, read_metrics
 from ._dts import read_dts
 from ._toy import toy